Set up articles page
Includes updates to itemlist component
